# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "Lavarel Study Swagger OpenApi description",
        "contact": {
            "email": "matheussouzajose97@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server Lavarel Study"
        }
    ],
    "paths": {
...
        "/email/verification-notification": {
            "post": {
                "tags": [
                    "Email"
                ],
                "summary": "Send email for confirmation user",
                "description": "Send email for confirmation user",
                "operationId": "verification-notification",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/email/verify-email/{id}/{hash}": {
            "get": {
                "tags": [
                    "Email"
                ],
                "summary": "Confirmation user",
                "description": "Confirmation user",
                "operationId": "verify-email",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "hash",
                        "in": "path",
                        "description": "hash",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmailResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/password/forgot-password": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Forgot password",
                "description": "Forgot password",
                "operationId": "forgot-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PasswordResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessablePassword"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/password/reset-password": {
            "post": {
                "tags": [
                    "Password"
                ],
                "summary": "Forgot password",
                "description": "Reset password",
                "operationId": "reset-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PasswordResetRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PasswordResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessablePassword"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Error": {
                "title": "Error Resource",
                "description": "Error resource",
                "properties": {
                    "error": {
                        "title": "error",
                        "description": "Error",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ErrorResource"
                }
            },
...
            "PasswordRequest": {
                "title": "Password user request",
                "description": "Password user request body data",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "email@gmail.com.br"
                    }
                },
                "type": "object"
            },
            "PasswordResetRequest": {
                "title": "Password user request",
                "description": "Password user request body data",
                "required": [
                    "email",
                    "password",
                    "password_confirmation",
                    "token"
                ],
                "properties": {
                    "email": {
                        "title": "email",
                        "description": "Email of the user",
                        "type": "string",
                        "example": "matheus@gmail.com"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the user",
                        "type": "string",
                        "example": "123456789"
                    },
                    "password_confirmation": {
                        "title": "password_confirmation",
                        "description": "Confirmation password of the user",
                        "type": "string",
                        "example": "123456789"
                    },
                    "token": {
                        "title": "token",
                        "description": "Token for reset password",
                        "type": "string",
                        "example": "441c6dbacfebdc913f5302cbc7212dde99d6aa41c271aafe9260f18256c2edf0"
                    }
                },
                "type": "object"
            },
...
            "EmailResource": {
                "title": "Email Resource",
                "description": "Email resource",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "EmailResource"
                }
            },
...
            "PasswordResource": {
                "title": "PasswordResource user",
                "description": "PasswordResource Content",
                "properties": {
                    "status": {
                        "title": "status",
                        "description": "Status Message",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "PasswordResource"
                }
            },
            "UnprocessablePassword": {
                "title": "Unprocessable",
                "description": "Unprocessable Content",
                "properties": {
                    "message": {
                        "title": "message",
                        "description": "Error Message",
                        "type": "string"
                    },
                    "errors": {
                        "title": "errors",
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Unprocessable"
                }
            }
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}